+++ Source/Sitecore.FakeDb/Pipelines/AddDbItem/EnsureIsChild.cs
@@ -1,0 +1,16 @@
+﻿namespace Sitecore.FakeDb.Pipelines.AddDbItem
+{
+  public class EnsureIsChild
+  {
+    public virtual void Process(AddDbItemArgs args)
+    {
+      var item = args.DbItem;
+      var dataStorage = args.DataStorage;
+
+      if (!dataStorage.GetFakeItem(item.ParentID).Children.Contains(item))
+      {
+        dataStorage.GetFakeItem(item.ParentID).Children.Add(item);
+      }
+    }
+  }
+}