+++ Source/Sitecore.FakeDb/Pipelines/InitFakeDb/InitAuthorizationProvider.cs
@@ -1,8 +1,6 @@
 ﻿namespace Sitecore.FakeDb.Pipelines.InitFakeDb
 {
-  using Sitecore.FakeDb.Security.AccessControl;
   using Sitecore.Security.AccessControl;
-  using System.Collections.Generic;
 
   public class InitAuthorizationProvider : InitDbProcessor
   {