+++ Source/Sitecore.FakeDb/Pipelines/InitFakeDb/InitGlobals.cs
@@ -1,10 +1,10 @@
 ﻿namespace Sitecore.FakeDb.Pipelines.InitFakeDb
 {
-  public class InitGlobals: InitDbProcessor
+  public class InitGlobals : InitDbProcessor
   {
     public override void Process(InitDbArgs args)
     {
-      Sitecore.Globals.Load();
+      Globals.Load();
     }
   }
 }